requirements.txt
deep_translator
openai-whisper
datetime
gradio